Fix issue with autoupdater
Autoupdater needs to be shut off on logout, as the user may login on the admin port.
This can leave it running. It also needs to be destroyed when disabled. Lastly, the
ontology switched events sometimes arrive before the logout event, causing a new autoupdater
to attempt starting. It runs into a null ontology, so a check is added to ensure an ontology
is present before starting.